+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This project adheres to [Semantic Versioning](https://semver.org)
 * Screens/index: utilize 100% of terminal width.
 * Screens/index: possible line overflow.
 * Screens/log: segmentation fault when toggling the log in a detached state.
+* Screens/log: scroll "jitter" with multi-line entries. #46
 
 ### [v0.7.0] - 2019-02-28