+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](https://semver.org)
 
 ## [Unreleased]
 
+## [v0.8.0] - 2019-05-26
+
 ### Added
 - A signal hander. Send `SIGINT` (^C) to prematurely terminate bookwyrm.
 
@@ -50,6 +52,7 @@ This project adheres to [Semantic Versioning](https://semver.org)
 
 * Initial (usable) release
 
-[Unreleased]: https://github.com/Tmplt/bookwyrm/compare/v0.7.0...HEAD
+[Unreleased]: https://github.com/Tmplt/bookwyrm/compare/v0.8.0...HEAD
+[v0.8.0]: https://github.com/Tmplt/bookwyrm/compare/v0.7.0...v0.8.0
 [v0.7.0]: https://github.com/Tmplt/bookwyrm/compare/v0.6.1...v0.7.0
 [v0.6.1]: https://github.com/Tmplt/bookwyrm/compare/v0.6.0...v0.6.1